fix(gpui-client): address review feedback — move Mode to models, fix URL bug, cleanup deps
- Move Mode enum from server crate to models crate so gpui-client
  doesn't pull in the heavyweight server dependency. server::Mode
  re-exports from models for backward compatibility.
- Fix URL trailing-slash bug in ApiClient::url() with trim_end_matches
- Change flush_merge_queue to use post_empty (no request body needed)
- Remove unused deps: serde_json, chrono, server
