+++ crates/gpui-client/Cargo.toml
@@ -9,13 +9,10 @@ description = "HTTP API client for the Tasks platform GPUI frontend"
 [dependencies]
 # Workspace dependencies
 serde.workspace = true
-serde_json.workspace = true
 tokio.workspace = true
-chrono.workspace = true
 thiserror.workspace = true
 reqwest.workspace = true
 
 # Internal crates
 models = { path = "../models" }
 events = { path = "../events" }
-server = { path = "../server" }
