+++ crates/models/Cargo.toml
@@ -9,3 +9,5 @@ rust-version.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 chrono.workspace = true
+thiserror.workspace = true
+events = { path = "../events" }
